% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bootstrapPath.R
\name{bootstrapPath}
\alias{bootstrapPath}
\title{Perform post hoc bootstrap analysis on pathway over-resentation}
\usage{
bootstrapPath(pathway.id, geneResult = NULL, proteinResult = NULL,
  metaboliteResult = NULL, geneCutoff = NULL, proteinCutoff = NULL,
  metaboliteCutoff = NULL, seed = 1990, numSims = 10000)
}
\arguments{
\item{pathway.id}{KEGG pathway name to perform bootstrap on.  ("hsa04727")}

\item{geneResult}{Dataframe with transcript IDs as rownames, transcript signifigance values (between 0 and 1) as the first column, and optionally transcript fold change values as the second column. Defaults to NULL}

\item{proteinResult}{Dataframe with protein accession numbers as rownames, protein signifigance values (between 0 and 1) as the first column, and optionally protein fold change values as the second column. Defaults to NULL}

\item{metaboliteResult}{Dataframe with metabolite IDs as rownames, metabolite signifigance values (between 0 and 1) as the first column, and optionally metabolite fold change values as the second column. Defaults to NULL}

\item{geneCutoff}{Significance cutoff to use for transcript analysis. Defaults to NULL}

\item{proteinCutoff}{Significance cutoff to use for protein analysis.  Defaults to NULL}

\item{metaboliteCutoff}{Significance cutoff to use for metabolite analysis. Defaults to NULL}

\item{seed}{Random number seed for reproducibility. Defaults to 1990}

\item{numSims}{Integer indicating the number of bootstrap simulations to perform.  Defaults to 10000}
}
\value{
Returns an histogram with pathway representation frequencies from bootstrap analysis.
}
\description{
This takes dataframes with rna and metabolite info and performs simulated bootstrap analysis of pathway over-representation.
}
\note{
bootstrapPath is powered by the following open source databases.  Commercial use and/or redistribution may restricted.  Please see respective terms of use pages and citations for more details.

KEGG

Terms of Use: http://www.kegg.jp/kegg/legal.html

Citations:

Kanehisa, M., Goto, S., Sato, Y., Kawashima, M., Furumichi, M., and Tanabe, M.  Data, information, knowledge and principle: back to metabolism in KEGG. Nucleic Acids Res. 42, D199-D205 (2014).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. Nucleic Acids Res. 28, 27-30 (2000).

SMPDB

Terms of Use: http://smpdb.ca/about

Citations:

Wishart DS, Frolkis A, Knox C, et al. SMPDB: The Small Molecule Pathway Database. Nucleic Acids Res. 2010 Jan;38(Database issue):D480-7.

Jewison T, Su Y, Disfany FM, et al. SMPDB 2.0: Big Improvements to the Small Molecule Pathway Database Nucleic Acids Res. 2014 Jan;42(Database issue):D478-84.

Reactome

Terms of Use: http://www.reactome.org/pages/about/license-agreement/

Citations:

Croft et al. 2014 PMID: 24243840

Milacic et al. 2012 PMID:24213504
}
\examples{
data(kData)
data(rData)
bootstrapPath("hsa04727", geneResult=kData, metaboliteResult=rData,
     geneCutoff=0.1, metaboliteCutoff=0.219)
}
\keyword{IPA}

