% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DrugBank_GeneTarget.R
\docType{data}
\name{DrugBank_GeneTarget}
\alias{DrugBank_GeneTarget}
\title{Pathway Bank gene target info}
\format{A dataframe with 2361 rows and 4987 variables.  All variables are individual drugs.}
\description{
This data set contains all genes targets of DrugBank drugs
}
\note{
The following open source databases were used to compile this file.  Commercial use and/or redistribution may restricted.  Please see respective terms of use pages and citations for more details.

DrugBank:

Terms of Use: http://www.drugbank.ca/about

Citations:

DrugBank 4.0: shedding new light on drug metabolism. Law V, Knox C, Djoumbou Y, Jewison T, Guo AC, Liu Y, Maciejewski A, Arndt D, Wilson M, Neveu V, Tang A, Gabriel G, Ly C, Adamjee S, Dame ZT, Han B, Zhou Y, Wishart DS.Nucleic Acids Res. 2014 Jan 1;42(1):D1091-7. PubMed ID: 24203711

DrugBank 3.0: a comprehensive resource for 'omics' research on drugs. Knox C, Law V, Jewison T, Liu P, Ly S, Frolkis A, Pon A, Banco K, Mak C, Neveu V, Djoumbou Y, Eisner R, Guo AC, Wishart DS.Nucleic Acids Res. 2011 Jan;39(Database issue):D1035-41.  PubMed ID: 21059682

DrugBank: a knowledgebase for drugs, drug actions and drug targets. Wishart DS, Knox C, Guo AC, Cheng D, Shrivastava S, Tzur D, Gautam B, Hassanali M.Nucleic Acids Res. 2008 Jan;36(Database issue):D901-6. PubMed ID: 18048412

DrugBank: a comprehensive resource for in silico drug discovery and exploration. Wishart DS, Knox C, Guo AC, Shrivastava S, Hassanali M, Stothard P, Chang Z, Woolsey J.Nucleic Acids Res. 2006 Jan 1;34(Database issue):D668-72.  PubMed ID: 16381955
}
\keyword{data}

