% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ipa.R
\name{ipa}
\alias{ipa}
\title{This takes dataframes with rna and metabolite info and performs pathway analysis}
\usage{
ipa(geneResult = NULL, proteinResult = NULL, metaboliteResult = NULL,
  combine = FALSE, method = c("fisher.exact", "EASE", "mean.significance",
  "hypergeometric"), geneCutoff = NULL, proteinCutoff = NULL,
  metaboliteCutoff = NULL, pathwayType = c("All", "KEGG", "SMPDB",
  "Reactome"), calculateFoldChange = TRUE, countThreshold = 2,
  p.adjustMethod = "BH")
}
\arguments{
\item{geneResult}{Dataframe with transcript IDs as rownames, transcript signifigance values (between 0 and 1) as the first column, and optionally transcript fold change values as the second column. Defaults to NULL}

\item{proteinResult}{Dataframe with Uniprot Accession IDs as rownames, Protein signifigance values (between 0 and 1) as the first column, and optionally protein fold change values as the second column. Defaults to NULL}

\item{metaboliteResult}{Dataframe with metabolite IDs as rownames, metabolite signifigance values (between 0 and 1) as the first column, and optionally metabolite fold change values as the second column. Defaults to NULL}

\item{combine}{Logical indicated whether to perform combined transcirpt and metabolite analysis.  Defaults to FALSE}

\item{method}{Method of pathway analysis to perform. Options include "fisher.exact",  "EASE", "mean.significance", and "hypergeometric.test". Defaults to "fisher.exact.}

\item{geneCutoff}{Significance cutoff to use for transcript analysis}

\item{proteinCutoff}{Significance cutoff to use for protein analysis}

\item{metaboliteCutoff}{Significance cutoff to use for metabolite analysis}

\item{pathwayType}{Pathway database to use for pathway analysis.  Options include "KEGG", "SMPDB", "Reactome", or "All".  Defaults to "All"}

\item{calculateFoldChange}{Logical indicating whether to perform cumulative fold change calculations.  If true, must have a second column with fold change numbers in geneResult and/or metaboliteResult input.  Defaults to TRUE.}

\item{countThreshold}{Integer indicating the minimum number of pathway members to require for a valid result. Defaults to 2.}

\item{p.adjustMethod}{Character indicating the type of multiple hypothesis correction to perform.  Defaults to "BH". See ?p.adjust() for details}
}
\value{
Returns a dataframe with pathway FDR corrected pathway significance and cumulative fold change for each pathway.
}
\description{
This takes dataframes with rna and metabolite info and performs pathway analysis
}
\note{
ipa is powered by the following open source databases.  Commercial use and/or redistribution may restricted.  Please see respective terms of use pages and citations for more details.

KEGG

Terms of Use: http://www.kegg.jp/kegg/legal.html

Citations:

Kanehisa, M., Goto, S., Sato, Y., Kawashima, M., Furumichi, M., and Tanabe, M.  Data, information, knowledge and principle: back to metabolism in KEGG. Nucleic Acids Res. 42, D199-D205 (2014).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. Nucleic Acids Res. 28, 27-30 (2000).

SMPDB

Terms of Use: http://smpdb.ca/about

Citations:

Wishart DS, Frolkis A, Knox C, et al. SMPDB: The Small Molecule Pathway Database. Nucleic Acids Res. 2010 Jan;38(Database issue):D480-7.

Jewison T, Su Y, Disfany FM, et al. SMPDB 2.0: Big Improvements to the Small Molecule Pathway Database Nucleic Acids Res. 2014 Jan;42(Database issue):D478-84.

Reactome

Terms of Use: http://www.reactome.org/pages/about/license-agreement/

Citations:

Croft et al. 2014 PMID: 24243840

Milacic et al. 2012 PMID:24213504
}
\examples{
## Parallel pathway analysis on individual datasets using Fisher's exact test.
data(kData)
data(rData)
ipa(geneResult=kData, metaboliteResult=rData, combine=FALSE, geneCutoff=0.1,
     metaboliteCutoff=0.219)

##Integrated pathway analysis using Fisher's exact test.
data(kData)
data(rData)
ipa(geneResult=kData, metaboliteResult=rData, combine=TRUE, geneCutoff=0.1,
     metaboliteCutoff=0.219)
}
\keyword{IPA}

