% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/kData_example.R
\docType{data}
\name{kData}
\alias{kData}
\title{Example Transcript Data}
\format{A dataframe with 14802 rows and 2 variables.}
\description{
This is an example transcript data file to be used for analysis.
}
\author{
Ryne Ramaker
}
\keyword{data}

