% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/KEGG_Chebi.R
\docType{data}
\name{KEGG_Chebi}
\alias{KEGG_Chebi}
\title{KEGG to Chebi ID conversion table}
\format{A dataframe with 16178 rows and 1 variables.}
\description{
This data set contains information on converting CHEBI IDs to KEGG IDs for the pathview function
}
\note{
KEGG

Terms of Use: http://www.kegg.jp/kegg/legal.html

Citations:

Kanehisa, M., Goto, S., Sato, Y., Kawashima, M., Furumichi, M., and Tanabe, M.  Data, information, knowledge and principle: back to metabolism in KEGG. Nucleic Acids Res. 42, D199-D205 (2014).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. Nucleic Acids Res. 28, 27-30 (2000).
}
\keyword{data}

