% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/manipulateIPA.R
\name{manipulateIPA}
\alias{manipulateIPA}
\title{Perform interactive pathway analysis on rna and metabolite datasets}
\usage{
manipulateIPA(geneResult = NULL, metaboliteResult = NULL, method,
  countThreshold = 2, p.adjustMethod = "BH", pathwayType = c("All",
  "KEGG", "SMPDB", "Reactome"))
}
\arguments{
\item{geneResult}{Dataframe with transcript IDs as rownames, transcript signifigance values (between 0 and 1) as the first column, and optionally transcript fold change values as the second column. Defaults to NULL}

\item{metaboliteResult}{Dataframe with metabolite IDs as rownames, metabolite signifigance values (between 0 and 1) as the first column, and optionally metabolite fold change values as the second column. Defaults to NULL}

\item{method}{method Method of pathway analysis to perform. Options include "fisher.exact",  "EASE", "mean.significance", and "hypergeometric.test". Defaults to "fisher.exact.}

\item{countThreshold}{Integer indicating the minimum number of pathway members to require for a valid result. Defaults to 2.}

\item{p.adjustMethod}{Character indicating the type of multiple hypothesis correction to perform.  See ?p.adjust() for details}

\item{pathwayType}{Pathway database to use for pathway analysis.  Options include "KEGG", "SMPDB", "Reactome", or "All".  Defaults to "All"}
}
\value{
Returns an interactive dataplot of top 5 significant pathways with ticker to adjust FDR thresholds
}
\description{
This takes dataframes with rna and metabolite info and performs interactive pathway analysis
}
\note{
ipa is powered by the following open source databases.  Commercial use and/or redistribution may restricted.  Please see respective terms of use pages and citations for more details.

KEGG

Terms of Use: http://www.kegg.jp/kegg/legal.html

Citations:

Kanehisa, M., Goto, S., Sato, Y., Kawashima, M., Furumichi, M., and Tanabe, M.  Data, information, knowledge and principle: back to metabolism in KEGG. Nucleic Acids Res. 42, D199-D205 (2014).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. Nucleic Acids Res. 28, 27-30 (2000).

SMPDB

Terms of Use: http://smpdb.ca/about

Citations:

Wishart DS, Frolkis A, Knox C, et al. SMPDB: The Small Molecule Pathway Database. Nucleic Acids Res. 2010 Jan;38(Database issue):D480-7.

Jewison T, Su Y, Disfany FM, et al. SMPDB 2.0: Big Improvements to the Small Molecule Pathway Database Nucleic Acids Res. 2014 Jan;42(Database issue):D478-84.

Reactome

Terms of Use: http://www.reactome.org/pages/about/license-agreement/

Citations:

Croft et al. 2014 PMID: 24243840

Milacic et al. 2012 PMID:24213504
}
\examples{
data(kData)
data(rData)
## Not run: manipulateIPA(kData, rData, method="fisher.exact") ## End(Not run)
}
\keyword{IPA}

