% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pathIndex.R
\docType{data}
\name{pathIndex}
\alias{pathIndex}
\title{Pathway ID to name conversion table}
\format{A dataframe with 2247 rows and 1 variable.}
\description{
This data set contains all pathway names for a given pathway ID
}
\note{
The following open source databases were used to compile this file.  Commercial use and/or redistribution may restricted.  Please see respective terms of use pages and citations for more details.

KEGG

Terms of Use: http://www.kegg.jp/kegg/legal.html

Citations:

Kanehisa, M., Goto, S., Sato, Y., Kawashima, M., Furumichi, M., and Tanabe, M.  Data, information, knowledge and principle: back to metabolism in KEGG. Nucleic Acids Res. 42, D199-D205 (2014).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. Nucleic Acids Res. 28, 27-30 (2000).

SMPDB

Terms of Use: http://smpdb.ca/about

Citations:

Wishart DS, Frolkis A, Knox C, et al. SMPDB: The Small Molecule Pathway Database. Nucleic Acids Res. 2010 Jan;38(Database issue):D480-7.

Jewison T, Su Y, Disfany FM, et al. SMPDB 2.0: Big Improvements to the Small Molecule Pathway Database Nucleic Acids Res. 2014 Jan;42(Database issue):D478-84.

Reactome

Terms of Use: http://www.reactome.org/pages/about/license-agreement/

Citations:

Croft et al. 2014 PMID: 24243840

Milacic et al. 2012 PMID:24213504
}
\keyword{data}

