% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotIPA.R
\name{plotIPA}
\alias{plotIPA}
\title{Plot result of ipa}
\usage{
plotIPA(x, cutoff = 10)
}
\arguments{
\item{x}{A dataframe output generated by the ipa function}

\item{cutoff}{The number of pathways to plot. Defaults to 10}
}
\value{
Barplot of ipa result
}
\description{
This function allows you to plot multiple independent ipa analyses.
}
\examples{
data(kData)
data(rData)
ipaResult<-ipa(geneResult=kData, metaboliteResult=rData, combine=TRUE,
     geneCutoff=0.1, metaboliteCutoff=0.219)
plotIPA(ipaResult)
}
\keyword{ipa}
\keyword{plot}

