% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rData_example.R
\docType{data}
\name{rData}
\alias{rData}
\title{Example Metabolite Data}
\format{A dataframe with 99 rows and 2 variables.}
\description{
This is an example metabolite data file to be used for analysis.
}
\author{
Ryne Ramaker
}
\keyword{data}

