% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/viewPath.R
\name{viewPath}
\alias{viewPath}
\title{View metabolite foldchanges overlaid on KEGG pathways}
\usage{
viewPath(geneResult = NULL, proteinResult = NULL, metaboliteResult = NULL,
  pathway.id, geneColors = c("green", "grey", "red"),
  metaboliteColors = c("blue", "grey", "yellow"), cutoff = 1)
}
\arguments{
\item{geneResult}{Dataframe with transcript IDs as rownames, transcript signifigance values (between 0 and 1) as the first column, and optionally transcript fold change values as the second column. Defaults to NULL}

\item{proteinResult}{Dataframe with protein accession numbers as rownames, protein signifigance values (between 0 and 1) as the first column, and optionally protein fold change values as the second column. Defaults to NULL}

\item{metaboliteResult}{Dataframe with metabolite IDs as rownames, metabolite signifigance values (between 0 and 1) as the first column, and optionally metabolite fold change values as the second column. Defaults to NULL}

\item{pathway.id}{The name of the KEGG pathway to plot (e.g. "hsa04727")}

\item{geneColors}{A vector of three colors to use to represent gene/protein fold changes. Defaults to c("green", "grey","red").}

\item{metaboliteColors}{A vector of three colors to use to represent metabolite fold changes.  Defaults to c("blue","grey","yellow")}

\item{cutoff}{The FDR to use for plotting. Defaults to 1}
}
\value{
Exports a png file to your working directory
}
\description{
This function is a wrapper for the pathview package which allows you to plot data on KEGG pathways
}
\note{
viewPath is powered by the following open source databases.  Commercial use and/or redistribution may restricted.  Please see respective terms of use pages and citations for more details.

#' KEGG

Terms of Use: http://www.kegg.jp/kegg/legal.html

Citations:

Kanehisa, M., Goto, S., Sato, Y., Kawashima, M., Furumichi, M., and Tanabe, M.  Data, information, knowledge and principle: back to metabolism in KEGG. Nucleic Acids Res. 42, D199-D205 (2014).

Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes and Genomes. Nucleic Acids Res. 28, 27-30 (2000).
}
\examples{
data(kData)
data(rData)
library(pathview)
data(demo.paths)
viewPath(geneResult=kData, metaboliteResult=rData, pathway.id=demo.paths$sel.paths[1])
}
\keyword{kegg}
\keyword{path}
\keyword{view}

